backend/requirements.txt
httpx
selectolax
fastapi
uvicorn
sqlalchemy
aiomysql
python-dotenv
apscheduler